+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Fixed:
+- Tree Export: Print tree to display correctly for subtrees / non-root nodes.
 
 ## [1.0.0] - 2025-09-26
 ### Added: